ai-service/requirements.txt
# ===========================================
# AI Backend Service Dependencies
# ===========================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Database
supabase==2.3.4
asyncpg==0.29.0
sqlalchemy==2.0.25

# AI/ML Libraries
transformers>=4.36.2
torch>=2.9.0
sentence-transformers>=2.2.2
spacy>=3.7.2
faiss-cpu>=1.7.4

# Vector Store
numpy==1.26.3

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
tenacity==8.2.3

# Logging
structlog==24.1.0
